feat: appdelegate 파일 분리
