+++ pennyway-client-iOS/pennyway-client-iOS/App/AppDelegate.swift
@@ -1,10 +1,6 @@
 import Firebase
 import FirebaseCore
 import FirebaseMessaging
-import GoogleSignIn
-import KakaoSDKAuth
-import KakaoSDKCommon
-import KakaoSDKUser
 import SwiftUI
 
 // MARK: - AppDelegate
@@ -57,7 +53,6 @@ extension AppDelegate: MessagingDelegate {
         if let fcmToken = fcmToken {
             KeychainHelper.saveFcmToken(fcmToken: fcmToken)
             Log.debug("fcmToken: \(fcmToken)")
-            Log.debug("load fcmToken: \(String(describing: KeychainHelper.loadFcmToken()))")
         }
     }
 }
@@ -99,49 +94,5 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
         Log.debug(userInfo)
 
         completionHandler()
-    }
-}
-
-// MARK: - pennyway_client_iOSApp
-
-@main
-struct pennyway_client_iOSApp: App {
-    @StateObject private var appViewModel = AppViewModel()
-    @UIApplicationDelegateAdaptor(AppDelegate.self) var delegate
-
-    init() {
-        let kakaoAppKey = Bundle.main.infoDictionary?["KakaoAppKey"] as! String
-        KakaoSDK.initSDK(appKey: kakaoAppKey, loggingEnable: false)
-    }
-
-    var body: some Scene {
-        WindowGroup {
-            if appViewModel.isLoggedIn || appViewModel.checkLoginState {
-                MainTabView()
-                    .onAppear(perform: UIApplication.shared.addTapGestureRecognizer)
-                    .onOpenURL { url in
-                        GIDSignIn.sharedInstance.handle(url)
-                    }
-                    .environmentObject(appViewModel)
-
-            } else {
-                if appViewModel.isSplashShown {
-                    LoginView()
-                        .onAppear(perform: UIApplication.shared.addTapGestureRecognizer)
-                        .onOpenURL { url in
-                            GIDSignIn.sharedInstance.handle(url)
-                        }
-                        .environmentObject(appViewModel)
-
-                } else {
-                    MainView()
-                        .onAppear(perform: UIApplication.shared.addTapGestureRecognizer)
-                        .onOpenURL { url in
-                            GIDSignIn.sharedInstance.handle(url)
-                        }
-                        .environmentObject(appViewModel)
-                }
-            }
-        }
     }
 }
