+++ pennyway-client-iOS/pennyway-client-iOS/App/pennyway_client_iOSApp.swift
@@ -1,0 +1,49 @@
+import GoogleSignIn
+import KakaoSDKAuth
+import KakaoSDKCommon
+import KakaoSDKUser
+import SwiftUI
+
+// MARK: - pennyway_client_iOSApp
+
+@main
+struct pennyway_client_iOSApp: App {
+    @StateObject private var appViewModel = AppViewModel()
+    @UIApplicationDelegateAdaptor(AppDelegate.self) var delegate
+
+    init() {
+        let kakaoAppKey = Bundle.main.infoDictionary?["KakaoAppKey"] as! String
+        KakaoSDK.initSDK(appKey: kakaoAppKey, loggingEnable: false)
+    }
+
+    var body: some Scene {
+        WindowGroup {
+            if appViewModel.isLoggedIn || appViewModel.checkLoginState {
+                MainTabView()
+                    .onAppear(perform: UIApplication.shared.addTapGestureRecognizer)
+                    .onOpenURL { url in
+                        GIDSignIn.sharedInstance.handle(url)
+                    }
+                    .environmentObject(appViewModel)
+
+            } else {
+                if appViewModel.isSplashShown {
+                    LoginView()
+                        .onAppear(perform: UIApplication.shared.addTapGestureRecognizer)
+                        .onOpenURL { url in
+                            GIDSignIn.sharedInstance.handle(url)
+                        }
+                        .environmentObject(appViewModel)
+
+                } else {
+                    MainView()
+                        .onAppear(perform: UIApplication.shared.addTapGestureRecognizer)
+                        .onOpenURL { url in
+                            GIDSignIn.sharedInstance.handle(url)
+                        }
+                        .environmentObject(appViewModel)
+                }
+            }
+        }
+    }
+}
